Mesa: [windows] Fix for RDP / Chrome Remote Desktop.
This updates the list of Mesa patches to include a new one that
actually fixes Mesa to work properly under Remote Terminal or
Chrome Remote Desktop.

The main issue was that the Mesa WGL backend was still relying
on host GDI32-provided GetPixelFormat() / SetPixelFormat() functions
even though this is not needed. The patch removes this dependency.

For context, see: https://android-review.googlesource.com/#/c/152706/

Change-Id: I49e1ecfb5708b6b61ea9c79c39372f01415e83f6