Ensure protobuf can be compiled on a machine without autotools.
This adds a patch to the protobuf sources that corresponds to
running the 'autogen.sh' script with Autotools installed. This
re-generates the configure / Makefile.in files, ensuring that
the sources can be later compiled on a machine where the
Autotools are not installed.

This is especially useful for the Darwin remote build :)

+ Fix location of gmock/gtest sub-directories.

Change-Id: I3f8643e7c416460b10f1be16f5b53f1ae84a52b1